Reworked scripts to build macOS binaries as either Universal, x86_64 or i386. Updated CMakeLists.txt to remove libuuid for macOS builds (not needed for Azure SDK on macOS).

+++ scripts/build_uuid.sh
@@ -47,35 +47,57 @@ if [[ "$PLATFORM" == "linux" ]]; then
     echo "Compiling UUID source"
     make install  || true
 elif [[ "$PLATFORM" == "darwin" ]]; then
-
-   export AL_OPTS="-I/usr/local/Cellar/automake/1.16.1_1/bin "
-   BUILD_DIR_64=$BUILD_DIR/uuid_64
-   make distclean > /dev/null 2>&1
-   if [ -e "Makefile" ] ; then 
+  export AL_OPTS="-I/usr/local/Cellar/automake/1.16.1_1/bin "
+  # Check to see if we are doing a universal build or not.
+  # If we are not doing a universal build, pick an arch to
+  # build
+  if [[ "$UNIVERSAL" == "true" ]]; then
+    echo "[INFO] Building Universal Binary"
+    BUILD_DIR_64=$BUILD_DIR/uuid_64
+    make distclean > /dev/null 2>&1
+    if [ -e "Makefile" ] ; then 
        make distclean > /dev/null || true
-   fi
-   echo "Generating UUID build system for 64 bit"
-   export CFLAGS="-fPIC -arch x86_64 -Xarch_x86_64 -DSIZEOF_LONG_INT=8 -mmacosx-version-min=10.12"
-   ./autogen.sh || true
-   ./configure --disable-all-programs --enable-libuuid --prefix=$BUILD_DIR_64  || true
-   echo "Compiling UUID source"
-   make install  || true
+    fi
+    echo "Generating UUID build system for 64 bit"
+    export CFLAGS="-fPIC -arch x86_64 -Xarch_x86_64 -DSIZEOF_LONG_INT=8 -mmacosx-version-min=10.12"
+    ./autogen.sh || true
+    ./configure --disable-all-programs --enable-libuuid --prefix=$BUILD_DIR_64  || true
+    echo "Compiling UUID source"
+    make install  || true
 
 
-   make distclean > /dev/null 2>&1
-   if [ -e "Makefile" ] ; then 
+    make distclean > /dev/null 2>&1
+    if [ -e "Makefile" ] ; then 
        make distclean > /dev/null || true
-   fi
-   echo "Generating UUID build system"
-   export CFLAGS="-fPIC -arch i386 -Xarch_i386 -DSIZEOF_LONG_INT=4 -mmacosx-version-min=10.12" 
-   ./autogen.sh || true
-   ./configure --disable-all-programs --enable-libuuid --prefix=$BUILD_DIR  || true
-   echo "Compiling UUID source for 32 bit"
-   make install  || true
-   mv $BUILD_DIR/lib/libuuid.a $BUILD_DIR/lib/libuuid_32.a
+    fi
+    echo "Generating UUID build system"
+    export CFLAGS="-fPIC -arch i386 -Xarch_i386 -DSIZEOF_LONG_INT=4 -mmacosx-version-min=10.12" 
+    ./autogen.sh || true
+    ./configure --disable-all-programs --enable-libuuid --prefix=$BUILD_DIR  || true
+    echo "Compiling UUID source for 32 bit"
+    make install  || true
+    mv $BUILD_DIR/lib/libuuid.a $BUILD_DIR/lib/libuuid_32.a
 
-   lipo -create $BUILD_DIR_64/lib/libuuid.a $BUILD_DIR/lib/libuuid_32.a -output $BUILD_DIR/lib/libuuid.a 
-   rm -rf $BUILD_DIR_64 
+    lipo -create $BUILD_DIR_64/lib/libuuid.a $BUILD_DIR/lib/libuuid_32.a -output $BUILD_DIR/lib/libuuid.a 
+    rm -rf $BUILD_DIR_64 
+  else
+    make distclean > /dev/null 2>&1
+    if [ -e "Makefile" ] ; then 
+       make distclean > /dev/null || true
+    fi
+    if [[ "$ARCH" == "x86" ]]; then
+      echo "Generating UUID build system for 32 bit"
+      export CFLAGS="-fPIC -arch i386 -Xarch_i386 -DSIZEOF_LONG_INT=4 -mmacosx-version-min=10.12" 
+    else
+      echo "Generating UUID build system for 64 bit"
+      export CFLAGS="-fPIC -arch x86_64 -Xarch_x86_64 -DSIZEOF_LONG_INT=8 -mmacosx-version-min=10.12"
+    fi
+    ./autogen.sh || true
+    ./configure --disable-all-programs --enable-libuuid --prefix=$BUILD_DIR  || true
+    echo "Compiling UUID source"
+    make install  || true
+  fi
+
 else
     echo "[ERROR] Unknown platform: $PLATFORM"
     exit 1
